requirements.txt
attrs==21.4.0
celery==5.3.1
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.3
clickclick==20.10.2
distlib==0.3.4
docopt==0.6.2
filelock==3.6.0
Flask==2.3.2
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.4.0
MarkupSafe==2.1.1
pipenv==2022.1.8
platformdirs==2.5.1
pyrsistent==0.18.1
python-dotenv==0.19.2
PyYAML==5.3.1
requests==2.27.1
six==1.16.0
structlog==20.1.0
urllib3==1.26.8
virtualenv==20.13.3
virtualenv-clone==0.5.7
Werkzeug==2.3.3
tenacity==8.0.1
cryptography==39.0.1
GitPython==3.1.27
py-healthcheck==1.10.1
dependency-injector==4.38.0
